Squashed 'external/PlayfieldEngine/' changes from 434b613..3835ddd
3835ddd Merge NoctLang
5f249b9 Squashed 'external/NoctLang/' changes from c7be64a..1e8adc0

git-subtree-dir: external/PlayfieldEngine
git-subtree-split: 3835dddf4641e286ffb1db85c423fc10e32591e3

+++ external/NoctLang/src/core/interpreter.c
@@ -104,8 +104,10 @@ static INLINE bool rt_get_u16(
                 return false;
         }
 
-        *val = (int)((uint32_t)func->bytecode[*pc] << 8) |
-                     (uint32_t)func->bytecode[*pc + 1];
+        *val = (int)(
+                (uint32_t)((uint32_t)func->bytecode[*pc] << 8) |
+                (uint32_t)func->bytecode[*pc + 1]
+               );
 
         *pc = *pc + 2;
 
@@ -124,8 +126,10 @@ static INLINE bool rt_get_tmpvar(
                 return false;
         }
 
-        *val = (int)((uint32_t)func->bytecode[*pc] << 8) |
-                     (uint32_t)func->bytecode[*pc + 1];
+        *val = (int)(
+                (uint32_t)((uint32_t)func->bytecode[*pc] << 8) |
+                (uint32_t)func->bytecode[*pc + 1]
+               );
         if ((uint32_t)*val >= func->tmpvar_size) {
                 rt_error(env, BROKEN_BYTECODE);
                 return false;
